Fix zero-flag parsing in trace checker

+++ test/trace_zero_flag.c
@@ -8,7 +8,7 @@ struct entry_expect trace[] = {
     EXPECT_END,
 };
 
-volatile int x = 0;
+volatile uint64_t x = 0;
 int
 main()
 {
